cpp01-ex01 done
