+++ hchang/.clang-format
@@ -1,0 +1,5 @@
+---
+BasedOnStyle: Google
+AlignConsecutiveAssignments: 'true'
+
+...
