+++ hchang/CPP01/ex00/srcs/newZombie.cpp
@@ -1,28 +1,19 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   megaphone.cpp                                      :+:      :+:    :+:   */
+/*   newZombie.cpp                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: hchang <hchang@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/10/20 16:18:04 by hchang            #+#    #+#             */
-/*   Updated: 2022/11/01 12:46:29 by hchang           ###   ########.fr       */
+/*   Created: 2022/10/28 12:00:30 by hchang            #+#    #+#             */
+/*   Updated: 2022/11/01 11:17:35 by hchang           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <iostream>
+#include "Zombie.hpp"
 
-int main(int argc, char** argv) {
-  std::string str;
-  int a;
-
-  if (argc == 1)
-    std::cout << "* LOUD AND UNBEARABLE FEEDBACK NOISE *";
-  else {
-    for (int i = 1; i < argc; i++) str += std::string(argv[i]);
-    for (int j = 0; j < (int)str.size(); j++) str[j] = std::toupper(str[j]);
-    std::cout << str;
-  }
-  std::cout << std::endl;
-  return (0);
+Zombie		*newZombie(std::string name)
+{   
+    Zombie *zombie = new Zombie(name);
+    return (zombie);
 }
