+++ hchang/CPP01/ex01/deps/Zombie.d
@@ -1,0 +1,1 @@
+objs/Zombie.o: srcs/Zombie.cpp incs/Zombie.hpp
