+++ hchang/CPP01/ex01/deps/main.d
@@ -1,0 +1,1 @@
+objs/main.o: srcs/main.cpp incs/Zombie.hpp
