+++ hchang/CPP01/ex01/deps/zomebieHorde.d
@@ -1,0 +1,1 @@
+objs/zomebieHorde.o: srcs/zomebieHorde.cpp incs/Zombie.hpp
